+++ oap-server/server-query-plugin/query-graphql-plugin/src/main/java/org/apache/skywalking/oap/query/graphql/resolver/EBPFProcessProfilingQuery.java
@@ -72,6 +72,6 @@ public class EBPFProcessProfilingQuery implements GraphQLQueryResolver {
     public EBPFProfilingAnalyzation analysisEBPFProfilingResult(List<String> scheduleIdList,
                                                                 List<EBPFProfilingAnalyzeTimeRange> timeRanges,
                                                                 EBPFProfilingAnalyzeAggregateType aggregateType) throws IOException {
-        return getQueryService().getEBPFProfilingAnalyzation(scheduleIdList, timeRanges);
+        return getQueryService().getEBPFProfilingAnalyzation(scheduleIdList, timeRanges, aggregateType);
     }
 }