+++ docs/en/changes/changes.md
@@ -63,6 +63,7 @@
 * [Breaking Change] SQL Database: remove SQL Database config `maxSizeOfArrayColumn` and `numOfSearchableValuesPerTag`.
 * [Breaking Change] SQL Database: move `Tags list` from `Segment`,`Logs`,`Alarms` to their additional table.
 * [Breaking Change] Remove `total` field in Trace, Log, Event, Browser log, and alarm list query.
+* Support `OFF_CPU` eBPF Profiling.
 
 #### UI